# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.AuthConfirmRequest:
    properties:
      otp:
        type: string
      phone:
        type: string
    required:
    - otp
    - phone
    type: object
  auth.AuthLoginRequest:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  auth.AuthLoginResponse:
    properties:
      token:
        $ref: '#/definitions/auth.TokenDTO'
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.AuthMeResponse:
    properties:
      user:
        additionalProperties: {}
        type: object
    type: object
  auth.AuthRefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.AuthRegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - first_name
    - password
    - phone
    type: object
  auth.AuthRegisterResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.GoogleAuthRequest:
    properties:
      id_token:
        type: string
    required:
    - id_token
    type: object
  auth.TokenDTO:
    properties:
      access:
        type: string
      refresh:
        type: string
    type: object
  auth.UserDTO:
    properties:
      balance:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
      template_balance:
        type: integer
      username:
        type: string
    type: object
  dto.BaseResponse:
    properties:
      data: {}
      error:
        type: string
      status:
        type: boolean
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/auth/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Confirm request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.TokenDTO'
              type: object
      summary: Confirm phone number
      tags:
      - auth
  /api/auth/google:
    post:
      consumes:
      - application/json
      parameters:
      - description: Google auth request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.GoogleAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthLoginResponse'
              type: object
      summary: Google authentication
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthLoginResponse'
              type: object
      summary: Login user
      tags:
      - auth
  /api/auth/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthMeResponse'
              type: object
      summary: Get user profile
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.TokenDTO'
              type: object
      summary: Refresh token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthRegisterResponse'
              type: object
      summary: Register user
      tags:
      - auth
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"